Fix zero-copy not working
Up to two completion entries can be generated for a single send_zc
submission. The first one has a `more` flag and the second one has a
`notif` flag that indicates that the buffer is no longer in use by the
kernel.

We were taking both CQEs into account so the `in_flight` counter would
become negative and the main loop would be stuck.